// src/types/mock-data.type.ts
export type MockData = {
  titles: string[];
  descriptions: string[];
  publishDates: number[];
  genres: string[];
  airDates: number[];
  ratings: number[];
  previewVideos: string[];
  videos: string[];
  actors: string[];
  directors: string[];
  durations: number[];
  commentsNumbers: number[];
  users: string[];
  posters: string[];
  backgroundImages: string[];
  backgroundColors: string[];
};